feat: Refactor client details components and implement ClientAgentPackageList for improved access management
