+++ src/features/amUI/common/ClientAgentPackageList/useClientAccessPackageActions.ts
@@ -7,19 +7,19 @@ import type {
   RemoveAgentAccessPackagesFn,
 } from '@/rtk/features/clientApi';
 
-type UseClientDetailsAccessPackageActionsParams = {
+type UseClientAccessPackageActionsParams = {
   fromPartyUuid?: string;
   actingPartyUuid?: string;
   addAgentAccessPackages: AddAgentAccessPackagesFn;
   removeAgentAccessPackages: RemoveAgentAccessPackagesFn;
 };
 
-export const useClientDetailsAccessPackageActions = ({
+export const useClientAccessPackageActions = ({
   fromPartyUuid,
   actingPartyUuid,
   addAgentAccessPackages,
   removeAgentAccessPackages,
-}: UseClientDetailsAccessPackageActionsParams) => {
+}: UseClientAccessPackageActionsParams) => {
   const { t } = useTranslation();
   const { openSnackbar } = useSnackbar();
